# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "REHAU Nea Smart 2 — Bridge API",
    "description": "REST + SSE access to a REHAU Nea Smart 2.0 base station.",
    "version": "0.1.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  },
  "paths": {
...
    "/api/v1/rooms": {
      "get": {
        "tags": [
          "rooms"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "string"
                      },
                      "zone": {
                        "type": "integer"
                      },
                      "name": {
                        "type": "string"
                      },
                      "temperature": {
                        "type": "number"
                      },
                      "humidity": {
                        "type": "number"
                      },
                      "setpointHeating": {
                        "type": "number"
                      },
                      "setpointCooling": {
                        "type": "number"
                      },
                      "mode": {
                        "type": "string",
                        "enum": [
                          "standby",
                          "normal",
                          "reduced",
                          "program",
                          "program_override"
                        ]
                      },
                      "programOverride": {
                        "type": "boolean"
                      },
                      "hasFan": {
                        "type": "boolean"
                      },
                      "hasFlap": {
                        "type": "boolean"
                      },
                      "hasLight": {
                        "type": "boolean"
                      },
                      "fan": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 4
                      },
                      "flap": {
                        "type": "integer",
                        "minimum": 0,
                        "maximum": 1
                      },
                      "light": {
                        "type": "boolean"
                      },
                      "calibrationTemp": {
                        "type": "number"
                      },
                      "calibrationHumidity": {
                        "type": "number"
                      },
                      "programDailyId": {
                        "type": "integer"
                      },
                      "programWeeklyId": {
                        "type": "integer"
                      },
                      "floor": {
                        "type": "string"
                      },
                      "meta": {
                        "type": "object",
                        "properties": {
                          "lastUpdatedAt": {
                            "type": "string",
                            "description": "ISO timestamp"
                          }
                        },
                        "required": [
                          "lastUpdatedAt"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "required": [
                      "id",
                      "zone",
                      "name",
                      "temperature",
                      "humidity",
                      "setpointHeating",
                      "setpointCooling",
                      "mode",
                      "programOverride",
                      "hasFan",
                      "hasFlap",
                      "hasLight",
                      "fan",
                      "flap",
                      "light",
                      "calibrationTemp",
                      "calibrationHumidity",
                      "programDailyId",
                      "programWeeklyId",
                      "floor",
                      "meta"
                    ],
                    "additionalProperties": false
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/{id}": {
      "get": {
        "tags": [
          "rooms"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "zone": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "setpointHeating": {
                      "type": "number"
                    },
                    "setpointCooling": {
                      "type": "number"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "standby",
                        "normal",
                        "reduced",
                        "program",
                        "program_override"
                      ]
                    },
                    "programOverride": {
                      "type": "boolean"
                    },
                    "hasFan": {
                      "type": "boolean"
                    },
                    "hasFlap": {
                      "type": "boolean"
                    },
                    "hasLight": {
                      "type": "boolean"
                    },
                    "fan": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 4
                    },
                    "flap": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "light": {
                      "type": "boolean"
                    },
                    "calibrationTemp": {
                      "type": "number"
                    },
                    "calibrationHumidity": {
                      "type": "number"
                    },
                    "programDailyId": {
                      "type": "integer"
                    },
                    "programWeeklyId": {
                      "type": "integer"
                    },
                    "floor": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "id",
                    "zone",
                    "name",
                    "temperature",
                    "humidity",
                    "setpointHeating",
                    "setpointCooling",
                    "mode",
                    "programOverride",
                    "hasFan",
                    "hasFlap",
                    "hasLight",
                    "fan",
                    "flap",
                    "light",
                    "calibrationTemp",
                    "calibrationHumidity",
                    "programDailyId",
                    "programWeeklyId",
                    "floor",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "rooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "setpoint": {
                    "type": "number",
                    "minimum": 5,
                    "maximum": 31
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "standby",
                      "normal",
                      "reduced",
                      "program",
                      "program_override"
                    ]
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "zone": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "setpointHeating": {
                      "type": "number"
                    },
                    "setpointCooling": {
                      "type": "number"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "standby",
                        "normal",
                        "reduced",
                        "program",
                        "program_override"
                      ]
                    },
                    "programOverride": {
                      "type": "boolean"
                    },
                    "hasFan": {
                      "type": "boolean"
                    },
                    "hasFlap": {
                      "type": "boolean"
                    },
                    "hasLight": {
                      "type": "boolean"
                    },
                    "fan": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 4
                    },
                    "flap": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "light": {
                      "type": "boolean"
                    },
                    "calibrationTemp": {
                      "type": "number"
                    },
                    "calibrationHumidity": {
                      "type": "number"
                    },
                    "programDailyId": {
                      "type": "integer"
                    },
                    "programWeeklyId": {
                      "type": "integer"
                    },
                    "floor": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "id",
                    "zone",
                    "name",
                    "temperature",
                    "humidity",
                    "setpointHeating",
                    "setpointCooling",
                    "mode",
                    "programOverride",
                    "hasFan",
                    "hasFlap",
                    "hasLight",
                    "fan",
                    "flap",
                    "light",
                    "calibrationTemp",
                    "calibrationHumidity",
                    "programDailyId",
                    "programWeeklyId",
                    "floor",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/{id}/setpoint": {
      "put": {
        "tags": [
          "rooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "value": {
                    "type": "number",
                    "minimum": 5,
                    "maximum": 31,
                    "multipleOf": 0.5
                  }
                },
                "required": [
                  "value"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "zone": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "setpointHeating": {
                      "type": "number"
                    },
                    "setpointCooling": {
                      "type": "number"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "standby",
                        "normal",
                        "reduced",
                        "program",
                        "program_override"
                      ]
                    },
                    "programOverride": {
                      "type": "boolean"
                    },
                    "hasFan": {
                      "type": "boolean"
                    },
                    "hasFlap": {
                      "type": "boolean"
                    },
                    "hasLight": {
                      "type": "boolean"
                    },
                    "fan": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 4
                    },
                    "flap": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "light": {
                      "type": "boolean"
                    },
                    "calibrationTemp": {
                      "type": "number"
                    },
                    "calibrationHumidity": {
                      "type": "number"
                    },
                    "programDailyId": {
                      "type": "integer"
                    },
                    "programWeeklyId": {
                      "type": "integer"
                    },
                    "floor": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "id",
                    "zone",
                    "name",
                    "temperature",
                    "humidity",
                    "setpointHeating",
                    "setpointCooling",
                    "mode",
                    "programOverride",
                    "hasFan",
                    "hasFlap",
                    "hasLight",
                    "fan",
                    "flap",
                    "light",
                    "calibrationTemp",
                    "calibrationHumidity",
                    "programDailyId",
                    "programWeeklyId",
                    "floor",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/rooms/{id}/mode": {
      "put": {
        "tags": [
          "rooms"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "mode": {
                    "type": "string",
                    "enum": [
                      "standby",
                      "normal",
                      "reduced",
                      "program",
                      "program_override"
                    ]
                  },
                  "setpoint": {
                    "type": "number",
                    "minimum": 5,
                    "maximum": 31
                  }
                },
                "required": [
                  "mode"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {
                      "type": "string"
                    },
                    "zone": {
                      "type": "integer"
                    },
                    "name": {
                      "type": "string"
                    },
                    "temperature": {
                      "type": "number"
                    },
                    "humidity": {
                      "type": "number"
                    },
                    "setpointHeating": {
                      "type": "number"
                    },
                    "setpointCooling": {
                      "type": "number"
                    },
                    "mode": {
                      "type": "string",
                      "enum": [
                        "standby",
                        "normal",
                        "reduced",
                        "program",
                        "program_override"
                      ]
                    },
                    "programOverride": {
                      "type": "boolean"
                    },
                    "hasFan": {
                      "type": "boolean"
                    },
                    "hasFlap": {
                      "type": "boolean"
                    },
                    "hasLight": {
                      "type": "boolean"
                    },
                    "fan": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 4
                    },
                    "flap": {
                      "type": "integer",
                      "minimum": 0,
                      "maximum": 1
                    },
                    "light": {
                      "type": "boolean"
                    },
                    "calibrationTemp": {
                      "type": "number"
                    },
                    "calibrationHumidity": {
                      "type": "number"
                    },
                    "programDailyId": {
                      "type": "integer"
                    },
                    "programWeeklyId": {
                      "type": "integer"
                    },
                    "floor": {
                      "type": "string"
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "lastUpdatedAt": {
                          "type": "string",
                          "description": "ISO timestamp"
                        }
                      },
                      "required": [
                        "lastUpdatedAt"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "id",
                    "zone",
                    "name",
                    "temperature",
                    "humidity",
                    "setpointHeating",
                    "setpointCooling",
                    "mode",
                    "programOverride",
                    "hasFan",
                    "hasFlap",
                    "hasLight",
                    "fan",
                    "flap",
                    "light",
                    "calibrationTemp",
                    "calibrationHumidity",
                    "programDailyId",
                    "programWeeklyId",
                    "floor",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string"
                    },
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ]
}